Move InstAnalysis in Utility
add const on Assembly and MCII/MRI
