+++ src/Utility/CMakeLists.txt
@@ -1,13 +1,14 @@
 # Add QBDI target
 set(SOURCES
-    "${CMAKE_CURRENT_LIST_DIR}/memory_ostream.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/Assembly.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/InstAnalysis.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/LogSys.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/Memory.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/Memory_${QBDI_PLATFORM}.cpp"
-    "${CMAKE_CURRENT_LIST_DIR}/System.cpp"
-    "${CMAKE_CURRENT_LIST_DIR}/LogSys.cpp"
-    "${CMAKE_CURRENT_LIST_DIR}/Version.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/String.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/System.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/Version.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/memory_ostream.cpp"
     )
 
 if(QBDI_PLATFORM_IOS)
